DogBreedClassification/requirements.txt
keras
numpy
streamlit
tensorflow
tensorflow_intel
opencv-python
pillow